requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
filelock==3.16.1
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
huggingface-hub==0.27.0
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
lightning-utilities==0.11.9
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.1
opencv-python==4.10.0.84
packaging==24.2
pillow==11.0.0
propcache==0.2.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.16.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.0.post0
PyYAML==6.0.2
QDarkStyle==3.2.3
QtPy==2.4.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.14.1
setuptools==75.6.0
six==1.17.0
sympy==1.13.1
threadpoolctl==3.5.0
timm==1.0.12
torch==2.5.1
torchmetrics==1.6.1
torchvision==0.20.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
